refactor(cloudflare-metrics): migrate call sites to canonical import paths
drop the compatibility re-exports that metrics.ts and graphql-client.ts
were forwarding through, and update every import to point at the file
that actually defines the symbol:

- Metric                           ./metric.js
- HeaderMetricsProvider            ./metric-providers.js
- InfluxMetricsProvider            ./metric-providers.js
- IMetricsProviderRepository       ./metric-providers.js
- takeLastFlushStats               ./flush-state.js
- __resetFlushStateForTests        ./flush-state.js
- CloudflareMetricsRepository      ./metrics.js
- buildBatchedAccountQuery         ./graphql-builders.js
- buildBatchedZoneQuery            ./graphql-builders.js
- buildDatasetSelection            ./graphql-builders.js
- buildFilterObject                ./graphql-builders.js
- groupErrorsByAlias               ./graphql-builders.js
- CloudflareGraphQLClient          ./graphql-client.js (transport only)
- __resetResourceCachesForTests    ./resource-cache.js

also rename __resetMetricsModuleStateForTests → __resetFlushStateForTests
at every call site and drop the legacy alias. collector.ts no longer
re-exports __resetCollectorCachesForTests either — test files import
the canonical __resetResourceCachesForTests.

no behaviour changes; 71 tests still pass.
